pyproject.toml
[project]
name = "start"
version = "0.1.0"
description = "Active Inference curriculum generator and visualization toolkit"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Active Inference Institute" }]
license = { text = "MIT" }
dependencies = [
  "openai>=1.30.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "scikit-learn>=1.4.0",
  "networkx>=3.2.1",
  "nltk>=3.8.1",
  "spacy>=3.7.2",
  "wordcloud>=1.9.3",
  "adjustText>=0.8",
  "scipy>=1.12.0",
  "GitPython>=3.1.43",
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
]

[tool.uv]
package = false

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "black>=24.4",
  "ruff>=0.5",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B"]
ignore = []


